apps/ai-workers/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.0
rembg==2.0.56
onnxruntime==1.18.0
Pillow==10.3.0
diffusers==0.30.0
transformers==4.43.1
accelerate==0.33.0
safetensors==0.4.3
realesrgan==0.3.0
basicsr==1.4.2
facexlib==0.3.0
